+++ CHANGELOG.md
@@ -5,6 +5,10 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [Unreleased]
+
+### Fixed
+- app exclusions on Windows OS
+
 ## [1.5.0] - 2021-04-07
 ### Added
 - Punjabi translations